fix split dispaly
