+++ k9db-visualizer/src/components/FlowComponent/ControlPanelComponent/ControlPanel.tsx
@@ -1,19 +1,13 @@
 import React from "react";
 import MyControlButton from "./MyControlButton";
 import { Controls } from "reactflow";
-import SchemaWindow from "./SchemaWindow";
 
-interface Props {
-  onSelectAction: (action: string) => void;
-}
 
-function ControlPanel({ onSelectAction }: Props) {
-  // const handlers = [func1, func2, ]
+function ControlPanel() {
   return (
     <Controls>
-      {/* <MyControlButton name = "s"/> */}
-      <SchemaWindow />
       <MyControlButton name="v" action={() => console.log("validate")} />
+      <MyControlButton name="s" action={() => console.log("schema")} />
     </Controls>
   );
 }
