+++ k9db-visualizer/src/components/FlowComponent/Flow.tsx
@@ -1,0 +1,41 @@
+import React, { useCallback } from "react";
+import "reactflow/dist/style.css";
+import ReactFlow, {
+  MiniMap,
+  Background,
+  useNodesState,
+  useEdgesState,
+  addEdge,
+} from "reactflow";
+import ControlPanel from "./ControlPanelComponent/ControlPanel";
+
+const initialNodes = [
+    { id: "1", position: { x: 0, y: 0 }, data: { label: "1" } },
+    { id: "2", position: { x: 0, y: 100 }, data: { label: "2" } },
+  ];
+  const initialEdges = [{ id: "e1-2", source: "1", target: "2" }];
+  
+const Flow = () => {
+    const [nodes, setNodes, onNodesChange] = useNodesState(initialNodes);
+    const [edges, setEdges, onEdgesChange] = useEdgesState(initialEdges);
+  
+    const onConnect = useCallback(
+      (params) => setEdges((eds) => addEdge(params, eds)),
+      [setEdges]
+    );
+  return (
+    <ReactFlow
+          nodes={nodes}
+          edges={edges}
+          onNodesChange={onNodesChange}
+          onEdgesChange={onEdgesChange}
+          onConnect={onConnect}
+        >
+          <ControlPanel />
+          <MiniMap />
+          <Background gap={12} size={1} />
+        </ReactFlow>
+  )
+}
+
+export default Flow
